sync with main repo

+++ Rubberduck.Parsing/Symbols/ReferencedDeclarationsCollector.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.Diagnostics.Eventing.Reader;
 using System.Runtime.InteropServices;
 using System.Runtime.InteropServices.ComTypes;
 using Microsoft.Vbe.Interop;
@@ -90,9 +88,9 @@ namespace Rubberduck.Parsing.Symbols
                 case VarEnum.VT_USERDEFINED:
                     unchecked
                     {
-                        var href = desc.lpValue.ToInt32();
+                        var href = desc.lpValue.ToInt64();
                         ITypeInfo refTypeInfo;
-                        info.GetRefTypeInfo(href, out refTypeInfo);
+                        info.GetRefTypeInfo((int)href, out refTypeInfo);
                         return GetTypeName(refTypeInfo);
                     }
                 case VarEnum.VT_CARRAY:
